Fix GL precision mismatch on u_WaterlineZ; centre themed sun/moon; theme 윤슬 colors
- beach_vert.glsl: explicit mediump qualifier on u_WaterlineZ so vertex/fragment
  precision match (fixes shader link error)
- TideWatchRenderer: centerLightInView flag zeroes lightDir X after computeLightDir
  so sun/moon sits on the screen centre axis for AUTUMN_LIGHT and SUMMER_DARK themes;
  lightColorOverride replaces lutLight post-tint for themed 윤슬 sparkle
- TideWatchView: AUTUMN_LIGHT → warm gold override (1.0, 0.62, 0.22);
  SUMMER_DARK → silver moonlight override (0.72, 0.80, 0.95); others cleared

+++ app/src/main/res/raw/beach_vert.glsl
@@ -1,7 +1,7 @@
 precision highp float;
 attribute vec3  a_Pos;
-uniform mat4   u_MVP;
-uniform float  u_WaterlineZ;
+uniform mat4          u_MVP;
+uniform mediump float u_WaterlineZ;
 varying vec3   v_World;
 
 void main() {
